use Pool.Put to check that Stmt.Reset is called
A statement can be a bit tricky to use correctly.
It is easy after reading a single row to forget to call Reset
afterwards. This can result in a connection holding a lock, which
appears as a difficult-to-debug multi-connection bug.
To avoid that: check for reset on Pool.Put.

For #17