sqlite: fix two incomplete interrupt conditions
According to the SQLite docs:

	Any new SQL statements that are started after the
	sqlite3_interrupt() call and before the running statements
	reaches zero are interrupted as if they had been running
	prior to the sqlite3_interrupt() call.

When we detect a closed interrupt and return false from Step without
resetting a statemnt, we leave open a statement. This means future
calls to SetInterrupt don't succeed, because of an outstanding sqlite
statement.

To fix that conditon, reset stmts when we interrupt Step without
calling into cgo. To fix the more general condition, reset any
open statements on call to SetInterrupt.

Finally, make sure Reset always resets, as discovered by Amos Wenger.

For #18
For #20