sqlite: disable shared cache by default
The shared cache works well but is probably not the right default.
It has much stricter locking semantics than you get from SQLite
with a WAL database, reducing performance in some important cases.

This appears to be an implementation limit of the shared cache,
not something fundamental to its design. If it ever gets WAL-like
locking it should be turned on by default.