sqlite, sqlitex: trace query execution
This commit creates a runtime/trace task for each query,
processing sqlite3_step calls in regions.

There are two things missing from this approach.

First is we do not trace the reading and writing of variables.
Most of these are extremely short events, so much so that
creating a region seems unreasonable.
But it is perfectly possible to read a 1GB []byte.
Missing that is unfortunate.

Second, it would be nice to create tasks for transactions.
That would require sqlitex.Save to have more hooks into
an sqlite.Conn's tracer.