Properly finalize statements when closing connections (#51)
If close_v2 is called on a connection with unfinalized statements it
enters a zombie status, waiting for the statements to be finalized. This
leaks resources and prevents the WAL file from being deleted (the latter
being how I found this out).

close_v2 is for languages with unordered but reliable finalizers anyway.