Fix sqlitex_pool bit not being set correctly
Attempt to fix a bug in sqlitex pools where flags aren't being
propagated to OpenConn calls correctly. Flags ends up being
bitwise-and-ed to zero unless the caller of sqlitex.Open explicitly
adds this to the flags (and even then, it ends up being the only bit
set). With this, zero is always passed to sqlite.OpenConn, which
produces desirable default flags for a file-based pool, but invalid
flags for a memory-based pool with a size over 1.

This is related to but does not fix #43 (since this only visibly
affected memory pools for me). That's a matter of using appropriate
defaults for in-memory databases. It does make it possible to work
around it by passing the SQLITE_OPEN_SHAREDCACHE flag to sqlitex.Open,
so it's a start.