Fix deadlock in sqlitex.Open on error
Fix a bug where, when incorrect flags are passed to sqlitex.Open, any
error, such as SQLITE_MISUSE, causes sqlitex.Open to take the allMu
lock twice (nested). The first time the lock is taken is when it opens
connections. While this lock is held, if sqlite.OpenConn returns an
error, sqlitex.Open calls Close, which attempts to take the lock
a second time. This deadlocks sqlitex.Open.

To fix that, the Close call has been moved to a deferred call, outside
of the allMu lock, that checks if sqlitex.Open returned an error. If it
did, it then calls Close to clean up after itself. This adds named
return values to sqlitex.Open so that the error is visible from the
deferred call. The pool return argument is only assigned by return so
that p is never nil.