Fix _db_config args to avoid LoadExtension panic
LoadExtension would panic because sqlite3_db_config expects two
additional args when called with op
SQLITE_DBCONFIG_ENABLE_LOAD_EXTENSION. The second argument is a pointer
to the set value which can be used as an additional check. This is not
used but the NULL pointer arg must still be supplied to avoid a seg
fault.

https://www.sqlite.org/c3ref/c_dbconfig_defensive.html#sqlitedbconfigenableloadextension

Fix #62