Avoid MISUSE in Stmt.Bind
If a Conn is interrupted then Conn.Prep will return a Stmt with
Stmt.stmt = nil. This causes SQLITE MISUSE errors to occur when any of
the Bind functions are called on the Stmt. This is confusing to users as
the error returned by Step is misleading and anyone using sqlite.Logger
will see misleading MISUSE messages. Now we check if Stmt.stmt is not
nil prior to using it in the Bind statements, which solves these issues
and makes the interrupt obvious when Stmt.Step is later called.

fix #86