Add back blob API
Removed ReadAt and WriteAt methods, as these methods imply being safe to
call in parallel.

While this is the most efficient implementation of Read and Write, they
unfortunately don't pass -d=checkptr and require a non-moving GC. I'll
move to a less performant, but more safe, implementation in a future
commit.